+++ pkg/cmd/zkc/debug/observer.go
@@ -1,0 +1,241 @@
+// Copyright Consensys Software Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
+// the License. You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+// an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+// specific language governing permissions and limitations under the License.
+//
+// SPDX-License-Identifier: Apache-2.0
+package debug
+
+import (
+	"fmt"
+	"strings"
+
+	"github.com/consensys/go-corset/pkg/schema/register"
+	"github.com/consensys/go-corset/pkg/trace"
+	"github.com/consensys/go-corset/pkg/zkc/vm/function"
+	"github.com/consensys/go-corset/pkg/zkc/vm/instruction"
+	"github.com/consensys/go-corset/pkg/zkc/vm/machine"
+	"github.com/consensys/go-corset/pkg/zkc/vm/word"
+)
+
+// TraceObserver prints a trace
+type TraceObserver[W word.Word[W]] struct {
+	depth       uint
+	fun         *function.Function[instruction.Instruction[W]]
+	uses        string
+	definitions []register.Id
+	regWidth    uint
+	LeftPane    uint
+	MidPane     uint
+}
+
+// PreExecution implementation for Observer interface
+func (p *TraceObserver[W]) PreExecution(machine *machine.Base[W]) {
+	var (
+		n = machine.Depth()
+	)
+	//
+	if n > 0 {
+		if n != p.depth {
+			fmt.Println()
+			p.enterFunction(machine)
+			fmt.Print(p.callStack(machine))
+			fmt.Println()
+		}
+		//
+		p.writeInstruction(machine)
+	}
+}
+
+// PostExecution implementation for Observer interface
+func (p *TraceObserver[W]) PostExecution(machine *machine.Base[W]) {
+	var (
+		n = machine.Depth()
+	)
+	//
+	if n > 0 {
+		if n == p.depth {
+			p.writeState(machine)
+		}
+
+		fmt.Println()
+	}
+}
+
+func (p *TraceObserver[W]) enterFunction(machine *machine.Base[W]) {
+	var (
+		n     = machine.Depth()
+		frame = machine.StackFrame(n - 1)
+	)
+	//
+	p.depth = n
+	p.fun = machine.Module(frame.Function()).(*function.Function[instruction.Instruction[W]])
+	p.uses = ""
+	p.definitions = nil
+	p.regWidth = 0
+	//
+	for _, r := range p.fun.Registers() {
+		p.regWidth = max(p.regWidth, uint(len(r.Name())))
+	}
+}
+
+func (p *TraceObserver[W]) writeInstruction(machine *machine.Base[W]) {
+	var (
+		frame   = machine.StackFrame(p.depth - 1)
+		insn    = decode(frame, p.fun)
+		pc      = frame.PC()
+		name    = trace.ModuleName{Name: p.fun.Name(), Multiplier: 1}
+		mapping = instruction.NewSystemMap(register.ArrayMap(name, p.fun.Registers()...), machine.Modules())
+		insnStr = insn.String(mapping)
+		builder strings.Builder
+	)
+	//
+	insnStr = fmt.Sprintf("[%02x.%02x] %s", pc.Macro(), pc.Micro(), insnStr)
+	//
+	fmt.Print(leftAligned(insnStr, p.LeftPane))
+	// write uses
+	for i, r := range insn.Uses() {
+		var (
+			ith  = frame.Load(r.Unwrap())
+			name = p.fun.Register(r).Name()
+		)
+		//
+		if i != 0 {
+			builder.WriteString("; ")
+		}
+		//
+		fmt.Fprintf(&builder, "%s==0x%s", name, ith.Text(16))
+	}
+	//
+	p.uses = builder.String()
+	p.definitions = insn.Definitions()
+}
+
+func (p *TraceObserver[W]) writeState(machine *machine.Base[W]) {
+	fmt.Print(rightAligned(p.defs(machine), p.MidPane))
+	//
+	if len(p.definitions) > 0 || p.uses != "" {
+		fmt.Printf(" ; %s", p.uses)
+	}
+}
+
+func (p *TraceObserver[W]) defs(machine *machine.Base[W]) string {
+	var (
+		n       = machine.Depth()
+		frame   = machine.StackFrame(n - 1)
+		builder strings.Builder
+	)
+	//
+	for i, r := range p.definitions {
+		var (
+			ith = frame.Load(r.Unwrap())
+			reg = p.fun.Register(r)
+		)
+		//
+		if i != 0 {
+			builder.WriteString("; ")
+		}
+		//
+		builder.WriteString(rightAligned(reg.Name(), p.regWidth))
+		builder.WriteString(":=0x")
+		builder.WriteString(ith.Text(16))
+	}
+	//
+	return builder.String()
+}
+
+func (p *TraceObserver[W]) callStack(machine *machine.Base[W]) string {
+	var builder strings.Builder
+	//
+	for i := uint(0); i < p.depth; i++ {
+		var (
+			ith = machine.StackFrame(i)
+			fun = machine.Module(ith.Function()).(*function.Function[instruction.Instruction[W]])
+		)
+		//
+		if i+1 == p.depth {
+			inputs := functionInputs(ith, fun)
+			fmt.Fprintf(&builder, "> %s(%s) ", fun.Name(), inputs)
+		} else {
+			fmt.Fprintf(&builder, "> %s ", fun.Name())
+		}
+	}
+	//
+	return builder.String()
+}
+
+func functionInputs[W word.Word[W]](frame machine.Frame[W], fun *function.Function[instruction.Instruction[W]]) string {
+	var builder strings.Builder
+
+	for i, r := range fun.Registers() {
+		var ith = frame.Load(uint(i))
+		//
+		if !r.IsInput() {
+			break
+		} else if i != 0 {
+			builder.WriteString(", ")
+		}
+		//
+		fmt.Fprintf(&builder, "%s=0x%s", r.Name(), ith.Text(16))
+	}
+
+	return builder.String()
+}
+
+func leftAligned(str string, width uint) string {
+	var (
+		n       = min(uint(len(str)), width)
+		builder strings.Builder
+	)
+	//
+	if uint(len(str)) > width {
+		str = str[:width-2]
+		fmt.Fprintf(&builder, "%s..", str)
+	} else {
+		builder.WriteString(str)
+	}
+	//
+	for i := n; i < width; i++ {
+		builder.WriteString(" ")
+	}
+	//
+	return builder.String()
+}
+
+func rightAligned(str string, width uint) string {
+	var (
+		n       = min(uint(len(str)), width)
+		builder strings.Builder
+	)
+	//
+	for i := n; i < width; i++ {
+		builder.WriteString(" ")
+	}
+	//
+	builder.WriteString(str)
+	//
+	return builder.String()
+}
+
+func decode[W word.Word[W]](frame machine.Frame[W],
+	fn *function.Function[instruction.Instruction[W]]) instruction.MicroInstruction[W] {
+	//
+	var (
+		pc   = frame.PC()
+		insn = fn.CodeAt(pc.Macro())
+	)
+	// nolint
+	if uInsn, ok := insn.(*instruction.Vector[W]); ok {
+		return uInsn.Codes[pc.Micro()]
+	} else if uInsn, ok := insn.(instruction.MicroInstruction[W]); ok {
+		return uInsn
+	}
+	//
+	panic("invalid micro instruction")
+}
